requirements-dev.txt
-r requirements.txt
django-debug-toolbar>=4.2
django-extensions>=3.2
